+++ packages/cloud-toolkit/src/services/bootstrap.ts
@@ -1,4 +1,3 @@
-import { Table } from '@forestadmin/datasource-sql';
 import AdmZip from 'adm-zip';
 import * as fs from 'fs';
 import * as fsP from 'fs/promises';
@@ -16,7 +15,7 @@ async function tryToClearBootstrap(paths: BootstrapPathManager): Promise<string 
       fsP.rm(paths.cloudCustomizer, { force: false, recursive: true }),
     ]);
   } catch (e) {
-    return ` \nPlease remove cloud-customizer folder and re-run bootstrap command.`;
+    return `\nPlease remove cloud-customizer folder and re-run bootstrap command.`;
   }
 }
 
@@ -25,29 +24,6 @@ async function generateDotEnv(envSecret: string, paths: BootstrapPathManager) {
   let replaced = envTemplate.replace('<FOREST_ENV_SECRET_TO_REPLACE>', envSecret);
   replaced = replaced.replace('<TOKEN_PATH_TO_REPLACE>', paths.home);
   await fsP.writeFile(paths.env, replaced);
-}
-
-async function generateHelloWorldExample(
-  collectionName: string,
-  dependency: string,
-  paths: BootstrapPathManager,
-) {
-  const template = await fsP.readFile(paths.helloWorldTemplate, 'utf-8');
-  let replaced = template.replace('<COLLECTION_NAME_TO_REPLACE>', collectionName);
-  replaced = replaced.replace('<DEPENDENCY_TO_REPLACE>', dependency);
-  await fsP.writeFile(paths.index, replaced);
-}
-
-function findPrimaryKeyAndCollectionName(introspection: Table[]): {
-  collectionName: string;
-  primaryKey: string;
-} | null {
-  for (const collection of introspection) {
-    const pk = collection.columns.find(column => column.primaryKey);
-    if (pk) return { collectionName: collection.name, primaryKey: pk.name };
-  }
-
-  return null;
 }
 
 export default async function bootstrap(
@@ -74,8 +50,6 @@ export default async function bootstrap(
     if (!fs.existsSync(paths.env)) await generateDotEnv(envSecret, paths);
 
     const introspection = await httpServer.getIntrospection();
-    const data = findPrimaryKeyAndCollectionName(introspection);
-    if (data) await generateHelloWorldExample(data.collectionName, data.primaryKey, paths);
 
     await updateTypings(introspection, paths);
   } catch (error) {