+++ packages/cloud-toolkit/src/types.ts
@@ -78,6 +78,7 @@ export type MakeCommands = {
   buildEventSubscriber: BuildEventSubscriber;
   buildHttpServer: BuildHttpServer;
   getEnvironmentVariables: () => Promise<EnvironmentVariables>;
+  getCurrentVersion: () => string;
   bootstrapPathManager: BootstrapPathManager;
   distPathManager: DistPathManager;
   logger: Logger;