+++ packages/cloud-toolkit/test/services/bootstrap.unit.test.ts
@@ -21,6 +21,32 @@ jest.mock('../../src/services/update-typings', () => {
 });
 
 describe('bootstrap', () => {
+  const setupMocks = (introspection: Table[]) => {
+    jest.resetAllMocks();
+    jest.spyOn(fs, 'existsSync').mockReturnValue(false);
+    const writeFileSpy = jest.spyOn(fsP, 'writeFile');
+    jest.spyOn(fsP, 'readFile').mockImplementation(async () => {
+      return `
+          env: <FOREST_ENV_SECRET_TO_REPLACE>
+          path: <TOKEN_PATH_TO_REPLACE>
+          `;
+    });
+    jest.spyOn(fs, 'createWriteStream').mockReturnValue({
+      on: (eventName: string, callback: () => unknown) => {
+        callback();
+      },
+    } as unknown as fs.WriteStream);
+    const httpServer = new HttpServer('', '', '');
+    const path = new BootstrapPathManager('/tmp', '/my/home/directory');
+    jest.spyOn(httpServer, 'getIntrospection').mockResolvedValue(introspection);
+    HttpServer.downloadCloudCustomizerTemplate = jest.fn();
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    jest.mocked(AdmZip).mockReturnValue({ extractAllTo: () => {} });
+
+    return { writeFileSpy, httpServer, path };
+  };
+
   describe('If the target directory exists', () => {
     it('should throw a business error', async () => {
       jest.spyOn(fs, 'existsSync').mockReturnValue(true);
@@ -47,27 +73,30 @@ describe('bootstrap', () => {
           new BusinessError('Bootstrap failed: Failed.'),
         );
       });
+
+      describe('If there is an error when trying to clear the bootstrap', () => {
+        it('should notify the client to clear the cloud-customizer folder', async () => {
+          jest.spyOn(fs, 'existsSync').mockReturnValue(false);
+          const httpServer = new HttpServer('', '', '');
+          const path = new BootstrapPathManager('', '');
+          HttpServer.downloadCloudCustomizerTemplate = jest
+            .fn()
+            .mockRejectedValue(new Error('Failed'));
+          // throw error when trying to clear
+          jest.spyOn(fsP, 'rm').mockRejectedValue(new Error('Failed'));
+
+          await expect(bootstrap('abc', httpServer, path)).rejects.toEqual(
+            new BusinessError(
+              // eslint-disable-next-line max-len
+              'Bootstrap failed: Failed.\nPlease remove cloud-customizer folder and re-run bootstrap command.',
+            ),
+          );
+        });
+      });
     });
 
     describe('If no dependency call fails', () => {
       it('should run the bootstrap completely', async () => {
-        jest.spyOn(fs, 'existsSync').mockReturnValue(false);
-        const writeFileSpy = jest.spyOn(fsP, 'writeFile');
-        jest.spyOn(fsP, 'readFile').mockImplementation(async () => {
-          return `
-          env: <FOREST_ENV_SECRET_TO_REPLACE>
-          path: <TOKEN_PATH_TO_REPLACE>
-          collection: <COLLECTION_NAME_TO_REPLACE>
-          dep: <DEPENDENCY_TO_REPLACE>
-          `;
-        });
-        jest.spyOn(fs, 'createWriteStream').mockReturnValue({
-          on: (eventName: string, callback: () => unknown) => {
-            callback();
-          },
-        } as unknown as fs.WriteStream);
-        const httpServer = new HttpServer('', '', '');
-        const path = new BootstrapPathManager('/tmp', '/my/home/directory');
         const introspection: Table[] = [
           {
             name: 'towns',
@@ -87,27 +116,11 @@ describe('bootstrap', () => {
                 autoIncrement: true,
                 isLiteralDefaultValue: true,
               },
-              {
-                name: 'name',
-                type: {
-                  type: 'scalar',
-                  subType: 'STRING',
-                },
-                allowNull: false,
-                primaryKey: false,
-                constraints: [],
-                defaultValue: null,
-                autoIncrement: false,
-                isLiteralDefaultValue: false,
-              },
             ],
           },
         ];
-        jest.spyOn(httpServer, 'getIntrospection').mockResolvedValue(introspection);
-        HttpServer.downloadCloudCustomizerTemplate = jest.fn();
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        jest.mocked(AdmZip).mockReturnValue({ extractAllTo: () => {} });
+        const { writeFileSpy, httpServer, path } = setupMocks(introspection);
+
         const renameSpy = jest.spyOn(fsP, 'rename').mockResolvedValue();
         const rmSpy = jest.spyOn(fsP, 'rm').mockResolvedValue();
 
@@ -120,28 +133,15 @@ describe('bootstrap', () => {
         expect(renameSpy).toHaveBeenCalledWith('/tmp/cloud-customizer-main', 'cloud-customizer');
         expect(rmSpy).toHaveBeenCalledWith('/tmp/cloud-customizer.zip', { force: true });
 
-        expect(writeFileSpy).toHaveBeenCalledTimes(2);
+        expect(writeFileSpy).toHaveBeenCalledTimes(1);
         const firstCallArgs = writeFileSpy.mock.calls[0];
         expect(firstCallArgs[0]).toBe('cloud-customizer/.env');
         expect(firstCallArgs[1].toString().replace(/\s/g, '')).toBe(
           `
               env: abc
               path: /my/home/directory
-              collection: <COLLECTION_NAME_TO_REPLACE>
-              dep: <DEPENDENCY_TO_REPLACE>
               `.replace(/\s/g, ''),
         );
-        const secondCallArgs = writeFileSpy.mock.calls[1];
-        expect(secondCallArgs[0]).toBe('cloud-customizer/src/index.ts');
-        expect(secondCallArgs[1].toString().replace(/\s/g, '')).toBe(
-          `
-              env: <FOREST_ENV_SECRET_TO_REPLACE>
-              path: <TOKEN_PATH_TO_REPLACE>
-              collection: towns
-              dep: id
-              `.replace(/\s/g, ''),
-        );
-
         expect(updateTypings).toHaveBeenCalled();
         expect(updateTypings).toHaveBeenCalledWith(introspection, path);
       });